feat: 蓝盾对接权限中心RBAC #7794
Merge remote-tracking branch 'fc-github/issue-7794-master' into story-878402073-master

+++ src/backend/ci/ext/tencent/auth/biz-auth-tencent/src/main/kotlin/com/tencent/devops/auth/service/AuthPipelineIdService.kt
@@ -46,7 +46,7 @@ class AuthPipelineIdService @Autowired constructor(
         return if (idNumType) {
             if (resourceName.isNullOrEmpty()) {
                 val pipelineInfo = client.get(ServiceTXPipelineResource::class)
-                    .getPipelineInfobyId(resourceCode.toInt()).data
+                    .getPipelineInfobyId(resourceCode.toLong()).data
                     ?: throw PermissionForbiddenException(
                         message = MessageCodeUtil.getCodeMessage(
                             messageCode = CommonMessageCode.PERMISSION_DENIED,
